src/python_client/requirements.txt
Pillow==5.2.0
pkg-resources==0.0.0
redis==2.10.6
